Migrate to scale for blocks (#3377)
## Motivation
<!-- Please mention the issue fixed by this PR or detailed motivation -->
Partly Closes #3271 (for blocks)

## Changes
Applied scale codec to blocks

## Test Plan
<!-- Please specify how these changes were tested 
Unit tests

## DevOps Notes
<!-- Please uncheck these items as applicable to make DevOps aware of changes that may affect releases -->
- [x] This PR does not require configuration changes (e.g., environment variables, GitHub secrets, VM resources)
- [x] This PR does not affect public APIs
- [x] This PR does not rely on a new version of external services (PoET, elasticsearch, etc.)
- [x] This PR does not make changes to log messages (which monitoring infrastructure may rely on)

+++ common/types/transaction_header_scale.go
@@ -7,12 +7,12 @@ import (
 )
 
 func (t *Nonce) EncodeScale(enc *scale.Encoder) (total int, err error) {
-	if n, err := scale.EncodeCompact64(enc, t.Counter); err != nil {
+	if n, err := scale.EncodeCompact64(enc, uint64(t.Counter)); err != nil {
 		return total, err
 	} else {
 		total += n
 	}
-	if n, err := scale.EncodeCompact8(enc, t.Bitfield); err != nil {
+	if n, err := scale.EncodeCompact8(enc, uint8(t.Bitfield)); err != nil {
 		return total, err
 	} else {
 		total += n
@@ -25,13 +25,13 @@ func (t *Nonce) DecodeScale(dec *scale.Decoder) (total int, err error) {
 		return total, err
 	} else {
 		total += n
-		t.Counter = field
+		t.Counter = uint64(field)
 	}
 	if field, n, err := scale.DecodeCompact8(dec); err != nil {
 		return total, err
 	} else {
 		total += n
-		t.Bitfield = field
+		t.Bitfield = uint8(field)
 	}
 	return total, nil
 }
